DAOポータル再読み込み時にDAO名が表示されるよう修正
