+++ frontend/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useSubDAOList, useSubDAOData } from "./useSubDAO"
